+++ lib/src/kbasesearchengine/authorization/package-info.java
@@ -1,0 +1,5 @@
+/** Classes for managing authorization of users to access search data.
+ * @author gaprice@lbl.gov
+ *
+ */
+package kbasesearchengine.authorization;
