+++ lib/src/kbasesearchengine/main/SearchMethods.java
@@ -29,8 +29,8 @@ import kbasesearchengine.SearchTypesInput;
 import kbasesearchengine.SearchTypesOutput;
 import kbasesearchengine.SortingRule;
 import kbasesearchengine.TypeDescriptor;
+import kbasesearchengine.authorization.AccessGroupProvider;
 import kbasesearchengine.common.GUID;
-import kbasesearchengine.events.AccessGroupProvider;
 import kbasesearchengine.parse.KeywordParser;
 import kbasesearchengine.search.FoundHits;
 import kbasesearchengine.search.IndexingStorage;
